+++ lib/api-dto-mapper.js
@@ -1,6 +1,7 @@
 'use strict';
 
-const oas2MappingStrategy = require('./mapping-strategies/oas2');
+const oas2Strategy = require('./mapping-strategies/oas2');
+const oas3Strategy = require('./mapping-strategies/oas3');
 
 class ApiDto {
     #mappingStrategy;
@@ -29,7 +30,8 @@ class ApiDto {
 }
 
 const toDto = (spec, baseDir) => {
-    return new ApiDto(spec, baseDir, oas2MappingStrategy);
+    const mappingStrategy = spec.swagger ? oas2Strategy : oas3Strategy;
+    return new ApiDto(spec, baseDir, mappingStrategy);
 }
 
 module.exports = {
