+++ lib/mapping-strategies/oas3.js
@@ -4,7 +4,8 @@ const Path = require('path');
 const Utils = require('../../lib/utils');
 
 const getBasePath = function () {
-    return Utils.unsuffix(Utils.prefix(this.spec.basePath || '/', '/'), '/');
+    const server = this.spec.servers && this.spec.servers.length === 1 && this.spec.servers[0];
+    return Utils.unsuffix(Utils.prefix(server || '/', '/'), '/');
 }
 
 const getCustomAuthSchemes = function () {
@@ -16,7 +17,7 @@ const getCustomAuthSchemes = function () {
 }
 
 const getCustomAuthStrategies = function () {
-    const strategies = this.spec.securityDefinitions || {};
+    const strategies = this.spec.components && this.spec.components.securitySchemes || {};
 
     return Object.entries(strategies)
         .filter(([, config]) => config['x-hapi-auth-strategy'])
@@ -37,6 +38,33 @@ const mapSecurity = function (security = []) {
     );
 }
 
+const getRequestMediaTypes = function (operation) {
+    return operation.requestBody &&
+        operation.requestBody.content &&
+        Object.keys(operation.requestBody.content);
+}
+
+const mapRequestBodyToParameter = function (operation) {
+    const mediaTypes = getRequestMediaTypes(operation);
+    const schema = mediaTypes &&
+        mediaTypes.length &&
+        operation.requestBody.content[mediaTypes[0]].schema;
+
+    return schema ? [{ name: 'payload', in: 'body', schema }] : [];
+}
+
+const mapResponses = function (responses = {}) {
+    return Object.entries(responses).reduce((acc, [statusCode, response]) => {
+        const mappedResponse = { description: response.description };
+        if (response.content) {
+            const contentType = Object.keys(response.content)[0];
+            mappedResponse.schema = response.content[contentType].schema;
+        }
+        acc[statusCode] = mappedResponse;
+        return acc;
+    }, {});
+}
+
 const getOperations = function () {
     const globalSecurity = this.spec.security;
     const paths = this.spec.paths || {};
@@ -52,11 +80,15 @@ const getOperations = function () {
                 tags: operation.tags,
                 security: mapSecurity(operation.security || globalSecurity),
                 mediaTypes: {
-                    request: operation.consumes || this.spec.consumes,
+                    request: getRequestMediaTypes(operation),
                 },
-                parameters: [...(operations.parameters || []), ...(operation.parameters || [])],
+                parameters: [
+                    ...(operations.parameters || []),
+                    ...(operation.parameters || []),
+                    ...mapRequestBodyToParameter(operation),
+                ],
                 handler: operations['x-hapi-handler'] && Path.join(this.baseDir, operations['x-hapi-handler']),
-                responses: operation.responses,
+                responses: mapResponses(operation.responses),
                 customOptions: operation['x-hapi-options'],
             }))
     );
