+++ lib/routes.js
@@ -90,7 +90,7 @@ const create = async function ({ api, cors, vhost, handlers, extensions, outputv
             options.auth.access.scope = options.auth.access.scope || [];
             options.auth.access.scope.push(...secdef.scopes);
             options.auth.strategies = options.auth.strategies || [];
-            options.auth.strategies.push(secdef.schemeName);
+            options.auth.strategies.push(secdef.strategy);
 
             if (options.auth.access.scope.length === 0) {
                 options.auth.access.scope = false;
